+++ test/fixtures/CHANGELOG.md
@@ -1,9 +1,41 @@
-# CHANGELOG
+# Changelog
 
-## 0.1.2 - 2017-07-23
+All notable changes to this project will be documented in this file.
 
-* Tag without "v" prefix
-* Add a CHANGELOG.md file for testing
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 10.1.0-beta1 - 2021/02/08
+
+- This is the latest documented version in this changelog
+- The changelog module is tested against these lines
+- Be careful modifying this file
+
+## 10.1.0-alpha1 - 2021/01/10
+
+- This is a version with a prerelease in this changelog
+- The changelog module is tested against these lines
+- Be careful modifying this file
+
+### Fixed
+
+- trying with a subsection in a version note
+
+## 10.0.1 - 2020/12/31
+
+- End of year version
+
+## 10.0.0 - 11/04/2020
+
+- A
+- B
+- C
+
+## [9.10.1] - 2019-09-11
+
+- D
+- E
+- F
 
 ## v0.1.1 - 2017-07-22
 
@@ -13,5 +45,3 @@
 ## 0.1.0 - 2017-07-21
 
 * Very old version
-
-##